Update Modal logic
The pages can be requested as usual or in modal mode (`?modal`), which
results in just the page content.

There is a Router who switch between page Controllers, depending on the
uri and it's validity. These Controllers inherit\* from a basic class,
and may load specific resources, then they create the view with
apropriated template.

The PageApp view asks if it should output in modal mode or a full page,
the template to use and additional data to insert in the template. An
example is the data about the user, shown when _my profile_ is selected
in the aside menu.

\* The exception is the Login page, which is independent (because it
acts like an access control the the website)